Raul's feedback, refactored OnBlock and OnAttestation to its own file
