+++ beacon-chain/blockchain/forkchoice/service.go
@@ -3,14 +3,12 @@ package forkchoice
 import (
 	"bytes"
 	"context"
-	"fmt"
 	"sync"
 
 	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/go-ssz"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/state"
 	"github.com/prysmaticlabs/prysm/beacon-chain/db"
 	"github.com/prysmaticlabs/prysm/beacon-chain/db/filters"
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
@@ -244,218 +242,4 @@ func (s *Store) Head(ctx context.Context) ([]byte, error) {
 //    store.time = time
 func (s *Store) OnTick(t uint64) {
 	s.time = t
-}
-
-// OnBlock is called whenever a block is received. It runs state transition on the block and
-// update fork choice store struct.
-//
-// Spec pseudocode definition:
-//   def on_block(store: Store, block: BeaconBlock) -> None:
-//    # Make a copy of the state to avoid mutability issues
-//    assert block.parent_root in store.block_states
-//    pre_state = store.block_states[block.parent_root].copy()
-//    # Blocks cannot be in the future. If they are, their consideration must be delayed until the are in the past.
-//    assert store.time >= pre_state.genesis_time + block.slot * SECONDS_PER_SLOT
-//    # Add new block to the store
-//    store.blocks[signing_root(block)] = block
-//    # Check block is a descendant of the finalized block
-//    assert (
-//        get_ancestor(store, signing_root(block), store.blocks[store.finalized_checkpoint.root].slot) ==
-//        store.finalized_checkpoint.root
-//    )
-//    # Check that block is later than the finalized epoch slot
-//    assert block.slot > compute_start_slot_of_epoch(store.finalized_checkpoint.epoch)
-//    # Check the block is valid and compute the post-state
-//    state = state_transition(pre_state, block)
-//    # Add new state for this block to the store
-//    store.block_states[signing_root(block)] = state
-//
-//    # Update justified checkpoint
-//    if state.current_justified_checkpoint.epoch > store.justified_checkpoint.epoch:
-//        store.justified_checkpoint = state.current_justified_checkpoint
-//
-//    # Update finalized checkpoint
-//    if state.finalized_checkpoint.epoch > store.finalized_checkpoint.epoch:
-//        store.finalized_checkpoint = state.finalized_checkpoint
-func (s *Store) OnBlock(ctx context.Context, b *ethpb.BeaconBlock) error {
-	preState, err := s.db.State(ctx, bytesutil.ToBytes32(b.ParentRoot))
-	if err != nil {
-		return errors.Wrapf(err, "could not get pre state for slot %d", b.Slot)
-	}
-	if preState == nil {
-		return fmt.Errorf("pre state of slot %d does not exist", b.Slot)
-	}
-
-	// Blocks cannot be in the future. If they are, their consideration must be delayed until the are in the past.
-	slotTime := preState.GenesisTime + b.Slot*params.BeaconConfig().SecondsPerSlot
-	if slotTime > s.time {
-		return fmt.Errorf("could not process block from the future, slot time %d > current time %d", slotTime, s.time)
-	}
-
-	if err := s.db.SaveBlock(ctx, b); err != nil {
-		return errors.Wrapf(err, "could not save block from slot %d", b.Slot)
-	}
-
-	// Verify block is a descendent of a finalized block.
-	finalizedBlk, err := s.db.Block(ctx, bytesutil.ToBytes32(s.finalizedCheckpt.Root))
-	if err != nil || finalizedBlk == nil {
-		return errors.Wrap(err, "could not get finalized block")
-	}
-	root, err := ssz.SigningRoot(b)
-	if err != nil {
-		return errors.Wrapf(err, "could not get signing root of block %d", b.Slot)
-	}
-
-	bFinalizedRoot, err := s.ancestor(ctx, root[:], finalizedBlk.Slot)
-	if !bytes.Equal(bFinalizedRoot, s.finalizedCheckpt.Root) {
-		return fmt.Errorf("block from slot %d is not a descendent of the current finalized block", b.Slot)
-	}
-
-	// Verify block is later than the finalized epoch slot.
-	finalizedSlot := helpers.StartSlot(s.finalizedCheckpt.Epoch)
-	if finalizedSlot >= b.Slot {
-		return fmt.Errorf("block is equal or earlier than finalized block, slot %d < slot %d", b.Slot, finalizedSlot)
-	}
-
-	// Apply new state transition for the block to the store.
-	// Make block root as bad to reject in sync.
-	postState, err := state.ExecuteStateTransition(ctx, preState, b)
-	if err != nil {
-		if err := s.db.DeleteBlock(ctx, root); err != nil {
-			return errors.Wrap(err, "could not delete bad block from db")
-		}
-		return errors.Wrap(err, "could not execute state transition")
-	}
-
-	if err := s.db.SaveState(ctx, postState, root); err != nil {
-		return errors.Wrap(err, "could not save state")
-	}
-
-	// Update justified check point.
-	if postState.CurrentJustifiedCheckpoint.Epoch > s.justifiedCheckpt.Epoch {
-		s.justifiedCheckpt = postState.CurrentJustifiedCheckpoint
-	}
-
-	// Update finalized check point.
-	// Prune the block cache and helper caches on every new finalized epoch.
-	if postState.FinalizedCheckpoint.Epoch > s.finalizedCheckpt.Epoch {
-		helpers.ClearAllCaches()
-		s.finalizedCheckpt.Epoch = postState.FinalizedCheckpoint.Epoch
-	}
-
-	// Log epoch summary before the next epoch.
-	if helpers.IsEpochStart(postState.Slot) {
-		logEpochData(postState)
-	}
-
-	return nil
-}
-
-// OnAttestation is called whenever an attestation is received, it updates validators latest vote.
-//
-// Spec pseudocode definition:
-//   def on_attestation(store: Store, attestation: Attestation) -> None:
-//    target = attestation.data.target
-//
-//    # Cannot calculate the current shuffling if have not seen the target
-//    assert target.root in store.blocks
-//
-//    # Attestations cannot be from future epochs. If they are, delay consideration until the epoch arrives
-//    base_state = store.block_states[target.root].copy()
-//    assert store.time >= base_state.genesis_time + compute_start_slot_of_epoch(target.epoch) * SECONDS_PER_SLOT
-//
-//    # Store target checkpoint state if not yet seen
-//    if target not in store.checkpoint_states:
-//        process_slots(base_state, compute_start_slot_of_epoch(target.epoch))
-//        store.checkpoint_states[target] = base_state
-//    target_state = store.checkpoint_states[target]
-//
-//    # Attestations can only affect the fork choice of subsequent slots.
-//    # Delay consideration in the fork choice until their slot is in the past.
-//    attestation_slot = get_attestation_data_slot(target_state, attestation.data)
-//    assert store.time >= (attestation_slot + 1) * SECONDS_PER_SLOT
-//
-//    # Get state at the `target` to validate attestation and calculate the committees
-//    indexed_attestation = get_indexed_attestation(target_state, attestation)
-//    assert is_valid_indexed_attestation(target_state, indexed_attestation)
-//
-//    # Update latest messages
-//    for i in indexed_attestation.custody_bit_0_indices + indexed_attestation.custody_bit_1_indices:
-//        if i not in store.latest_messages or target.epoch > store.latest_messages[i].epoch:
-//            store.latest_messages[i] = LatestMessage(epoch=target.epoch, root=attestation.data.beacon_block_root)
-func (s *Store) OnAttestation(ctx context.Context, a *ethpb.Attestation) error {
-	tgt := a.Data.Target
-
-	// Verify beacon node has seen the target block before.
-	if !s.db.HasBlock(ctx, bytesutil.ToBytes32(tgt.Root)) {
-		return fmt.Errorf("target root %#x does not exist in db", bytesutil.Trunc(tgt.Root))
-	}
-
-	// Verify Attestations cannot be from future epochs.
-	// If they are, delay consideration until the epoch arrives
-	tgtSlot := helpers.StartSlot(tgt.Epoch)
-	baseState, err := s.db.State(ctx, bytesutil.ToBytes32(tgt.Root))
-	if err != nil {
-		return errors.Wrapf(err, "could not get pre state for slot %d", tgtSlot)
-	}
-	if baseState == nil {
-		return fmt.Errorf("pre state of target block %d does not exist", tgtSlot)
-	}
-
-	slotTime := baseState.GenesisTime + tgtSlot*params.BeaconConfig().SecondsPerSlot
-	if slotTime > s.time {
-		return fmt.Errorf("could not process attestation from the future epoch, time %d > time %d", slotTime, s.time)
-	}
-
-	// Store target checkpoint state if not yet seen.
-	h, err := hashutil.HashProto(tgt)
-	if err != nil {
-		return errors.Wrap(err, "could not hash justified checkpoint")
-	}
-	_, exists := s.checkptBlkRoot[h]
-	if !exists {
-		baseState, err = state.ProcessSlots(ctx, baseState, tgtSlot)
-		if err != nil {
-			return errors.Wrapf(err, "could not process slots up to %d", tgtSlot)
-		}
-		s.checkptBlkRoot[h] = bytesutil.ToBytes32(tgt.Root)
-	}
-
-	// Verify attestations can only affect the fork choice of subsequent slots.
-	aSlot, err := helpers.AttestationDataSlot(baseState, a.Data)
-	if err != nil {
-		return errors.Wrap(err, "could not get attestation slot")
-	}
-	slotTime = baseState.GenesisTime + (aSlot+1)*params.BeaconConfig().SecondsPerSlot
-	if slotTime > s.time {
-		return fmt.Errorf("could not process attestation for fork choice until inclusion delay, time %d > time %d", slotTime, s.time)
-	}
-
-	// Use the target state to to validate attestation and calculate the committees.
-	indexedAtt, err := blocks.ConvertToIndexed(baseState, a)
-	if err != nil {
-		return errors.Wrap(err, "could not convert attestation to indexed attestation")
-	}
-
-	if err := blocks.VerifyIndexedAttestation(baseState, indexedAtt); err != nil {
-		return errors.New("could not verify indexed attestation")
-	}
-
-	// Update every validator's latest vote.
-	for _, i := range append(indexedAtt.CustodyBit_0Indices, indexedAtt.CustodyBit_1Indices...) {
-		s.db.HasValidatorLatestVote(ctx, i)
-		vote, err := s.db.ValidatorLatestVote(ctx, i)
-		if err != nil {
-			return errors.Wrapf(err, "could not get latest vote for validator %d", i)
-		}
-		if !s.db.HasValidatorLatestVote(ctx, i) || tgt.Epoch > vote.Epoch {
-			if err := s.db.SaveValidatorLatestVote(ctx, i, &pb.ValidatorLatestVote{
-				Epoch: tgt.Epoch,
-				Root:  a.Data.BeaconBlockRoot,
-			}); err != nil {
-				return errors.Wrapf(err, "could not save latest vote for validator %d", i)
-			}
-		}
-	}
-	return nil
 }
